Adapt to change in McuCore: Using StatusCode rather than uint32_t as the type of codes stored in Status.

+++ src/device_types/observing_conditions/observing_conditions_adapter.cpp
@@ -227,7 +227,8 @@ bool ObservingConditionsAdapter::WriteDoubleOrSensorErrorResponse(
     mcucore::StatusOr<double> result, Print& out) {
   if (result.ok()) {
     return WriteResponse::DoubleResponse(request, result.value(), out);
-  } else if (result.status().code() == ErrorCodes::kNotImplemented) {
+  } else if (static_cast<int>(result.status().code()) ==
+             ErrorCodes::kNotImplemented) {
     return WriteSensorNotImpementedResponse(request, sensor_name, out);
   } else {
     return WriteResponse::StatusResponse(request, result.status(), out);
